modification du style gerer utilisateur (a continuer)

+++ code/controllers/c_gererUtilisateur.php
@@ -4,22 +4,22 @@
         && array_key_exists('type', $_POST)){
         require_once (PATH_MODELS.'AdminDAO.php');
         $dao = new AdminDAO(true, $_SESSION['login']);
-
-        if($_POST['type']=='ETUDIANT'){
-            $dao->updateStudent($_POST['prenom'],$_POST['nom'],$_POST['login']);
-            $dao->updateGroup($_POST['login'],$_POST['groups']);
+        if (isset($_POST["create"]) && $_POST["create"]=="créer"){
+            $dao->createCompte($_POST["login"],$_POST["mdp"],$_POST["prenom"],$_POST["nom"],$_POST["type"]);
         }
-        if($_POST['type']=='PROFESSEUR'){
-            $dao->updateTeacher($_POST['prenom'],$_POST['nom'],$_POST['login']);
-            $dao->deleteAffectation($_POST['login'],$_POST['type']);
-            for($i=1;$i<sizeof($_POST)-3;$i++){
-                $dao->assignerProf($_POST['login'],$_POST['module'.$i]);
+        else{
+            if($_POST['type']=='ETUDIANT'){
+                $dao->updateStudent($_POST['prenom'],$_POST['nom'],$_POST['login']);
+                $dao->updateGroup($_POST['login'],$_POST['groups']);
+            }
+            if($_POST['type']=='PROFESSEUR'){
+                $dao->updateTeacher($_POST['prenom'],$_POST['nom'],$_POST['login']);
+                $dao->deleteAffectation($_POST['login'],$_POST['type']);
+                for($i=1;$i<sizeof($_POST)-3;$i++){
+                    $dao->assignerProf($_POST['login'],$_POST['module'.$i]);
+                }
             }
         }
-
     }
 
 require_once (PATH_VIEWS.'gererUtilisateur.php');
-
-
-?>
